Rawr.DPSWarr: Fixed a crash issue with loading previously saved characters
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C36754

+++ Rawr.DPSWarr/CalculationOptionsDPSWarr.cs
@@ -13,7 +13,7 @@ namespace Rawr.DPSWarr
 		public string Filter = "All";
 		public string BossName = "Custom";
 		public int TargetLevel = 83;
-		public int TargetArmor = (int)StatConversion.NPC_ARMOR[83 - 80];
+		public float TargetArmor = StatConversion.NPC_ARMOR[83 - 80];
 		public float Duration = 300f;
 		public bool FuryStance = true;
 		public bool AllowFlooring = false;
